Reduce dynamic function binding from APIService from
